[ESSENTIAL]Extract helper methods to its own file, so will be reused and no more duplicated
